+++ src/main/java/com/example/alwayswin/controller/FigureController.java
@@ -55,6 +55,8 @@ public class FigureController {
     @ResponseBody
     @GetMapping("/product/figure/{fid}")
     CommonResult<Figure> getByFid(@PathVariable int fid){
+        if (fid <= 0)
+            return CommonResult.validateFailure();
         Figure figure = figureService.getFigureByFid(fid);
         if (figure == null) {
             return CommonResult.validateFailure();
@@ -65,6 +67,8 @@ public class FigureController {
     @ResponseBody
     @GetMapping("/product/{pid}/figures")
     CommonResult<List<Figure>> getByPid(@PathVariable int pid) {
+        if (pid <= 0)
+            return CommonResult.validateFailure();
 
         List<Figure> figureList = figureService.getFiguresByPid(pid);
         if (figureList == null) {
@@ -91,7 +95,9 @@ public class FigureController {
     @ResponseBody
     @PutMapping("/product/figure/edit/{fid}")
     CommonResult<Integer> editFigure(@RequestBody Map param, @PathVariable int fid){
-        int res = figureService.editFigure(fid, param);
+        if (fid < 0)
+            return CommonResult.validateFailure();
+        int res = figureService.updateFigure(fid, param);
         if (res == 1) {
             logger.info("Edit figure successfully");
             return CommonResult.success(res);
@@ -105,6 +111,9 @@ public class FigureController {
     @ResponseBody
     @DeleteMapping("/product/figure/delete/{fid}")
     CommonResult<Integer> deleteFigure(@PathVariable int fid){
+        if (fid < 0)
+            return CommonResult.validateFailure();
+
         int res = figureService.deleteFigure(fid);
         if (res == 1) {
             logger.info("Delete figure successfully");
